requirements.txt
certifi==2020.4.5.1
#cffi==1.14.5
chardet==3.0.4
cryptography==3.4.7
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2
django-allauth==0.44.0
django-appconf==1.0.4
django-crispy-forms==1.11.2
django-extensions==3.1.3
django-filebrowser-no-grappelli==3.8.0
django-imagekit==4.0.2
django-tinymce4-lite==1.8.0
idna==2.9
#jsmin==2.2.2
oauthlib==3.1.0
pilkit==2.0
#Pillow==8.2.0
pycparser==2.20
PyJWT==2.1.0
pyOpenSSL==19.1.0
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
